Add option to have non-square thumbnails and cover images
This commit adds support for non-square coverimages and
thumbnails. Two options have been added (square_thumbnails
and square_coverimage), which are True by default. Since
now the appropriate name for the dim_thumbnail is
dim_max_thumb, the options have been changed and a
DeprecationWarning is thrown when the settings file is
updated.

To correctly deal with EXIF rotation in SmartCrop.js, the
rotated file is saved to a temporary file which is cleaned
up after the crop has been created.