Fix issues with paths, extensions, date, and url_prefix
Paths on Windows use the backslash but the HTML path should use front
slash. When appending dimensions to the resized image filenames the
extension is now correctly placed only at the end. If no date is defined
in the album.yml then no date is emitted in the frontmatter. url_prefix
was meant to be optional but did not have correct defaults, now fixed.