Use cached_property decorator to simplify code
For some properties we already used a simple form of caching
of the return value, to make sure we don't run expensive
code too often. Now, we use an actual cached_property
decorator (taken from Django) to make this process simpler.
In addition, we now also cache opening of the original image
which makes the code run faster because the original image is
only opened once.