Add option to output albums as branches
Adds a .yml option: photos_as_resource_pages
If set to true, default behaviour will be changed to produce
- a directory `name` for the album instead of one `name.md` file
- an `_index.md` for the album inside this directory
- a `name.md` for every photo inside the same directory
Key properties of the photos will be saved in the front matter.

This change provides the basis for users who want to use keywords
to generate automatic collections of pages using Hugo's taxonomy
and have more finetuned control over the data being used to generate
the album.

To optimally use this new functionality, the user will need to
deploy new list templates that repicate the {{< wrap >}} and
produce the desired output.