Implement minimal EXIF and IPTC tag support
Use PIL's TAGS to deal with EXIF. Use
iptcinfo3 to deal with IPTC tags. Allows to
use these as fallback options for copyright
and caption of photos.

As a sample use-case also implements the
fallback for the photo album to be determined
from the set of photos in it.

Will not change the default behaviour without
adding the setting tag_map, exif or iptc to
the configuration file.